Split containsURL and isURL.
